订单中的数量从整数修改成小数

+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/purchase/returned/QueryPurchaseReturnBo.java
@@ -79,13 +79,13 @@ public class QueryPurchaseReturnBo extends BaseBo<PurchaseReturn> {
    * 采购数量
    */
   @ApiModelProperty("采购数量")
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
   @ApiModelProperty("赠品数量")
-  private Integer totalGiftNum;
+  private BigDecimal totalGiftNum;
 
   /**
    * 采购金额
