+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/retail/RetailProductBo.java
@@ -115,6 +115,6 @@ public class RetailProductBo extends BaseBo<RetailProductDto> {
     ProductStockService productStockService = ApplicationUtil.getBean(ProductStockService.class);
     ProductStock productStock = productStockService.getByProductIdAndScId(this.getProductId(),
         this.getScId());
-    this.stockNum = productStock == null ? 0 : productStock.getStockNum();
+    this.stockNum = productStock == null ? 0 : productStock.getStockNum().intValue();
   }
 }
