+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/sale/GetSaleOrderBo.java
@@ -367,7 +367,7 @@ public class GetSaleOrderBo extends BaseBo<SaleOrderFullDto> {
           ProductStockService.class);
       ProductStock productStock = productStockService.getByProductIdAndScId(
           this.getProductId(), this.getScId());
-      this.stockNum = productStock == null ? 0 : productStock.getStockNum();
+      this.stockNum = productStock == null ? 0 : productStock.getStockNum().intValue();
 
       if (StringUtil.isNotBlank(dto.getMainProductId())) {
         ProductService productService = ApplicationUtil.getBean(ProductService.class);
