+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/bo/stock/adjust/stock/StockAdjustSheetFullBo.java
@@ -304,7 +304,7 @@ public class StockAdjustSheetFullBo extends BaseBo<StockAdjustSheetFullDto> {
             ProductStockService.class);
         ProductStock productStock = productStockService.getByProductIdAndScId(dto.getProductId(),
             this.scId);
-        this.curStockNum = productStock == null ? 0 : productStock.getStockNum();
+        this.curStockNum = productStock == null ? 0 : productStock.getStockNum().intValue();
       }
     }
   }
