+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/dto/purchase/returned/PurchaseReturnFullDto.java
@@ -53,12 +53,12 @@ public class PurchaseReturnFullDto implements BaseDto, Serializable {
   /**
    * 商品数量
    */
-  private Integer totalNum;
+  private BigDecimal totalNum;
 
   /**
    * 赠品数量
    */
-  private Integer totalGiftNum;
+  private BigDecimal totalGiftNum;
 
   /**
    * 退货金额
@@ -138,7 +138,7 @@ public class PurchaseReturnFullDto implements BaseDto, Serializable {
     /**
      * 退货数量
      */
-    private Integer returnNum;
+    private BigDecimal returnNum;
 
     /**
      * 采购价
