+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/excel/purchase/PurchaseOrderImportModel.java
@@ -80,7 +80,7 @@ public class PurchaseOrderImportModel implements ExcelModel {
    */
   @ExcelRequired
   @ExcelProperty("采购数量")
-  private Integer purchaseNum;
+  private BigDecimal purchaseNum;
 
   /**
    * 是否赠品
