+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/impl/retail/RetailOutSheetServiceImpl.java
@@ -25,7 +25,7 @@ import com.lframework.starter.web.inner.components.timeline.ApproveReturnOrderTi
 import com.lframework.starter.web.inner.components.timeline.CreateOrderTimeLineBizType;
 import com.lframework.starter.web.inner.components.timeline.UpdateOrderTimeLineBizType;
 import com.lframework.starter.web.inner.dto.order.ApprovePassOrderDto;
-import com.lframework.starter.web.inner.dto.stock.ProductStockChangeDto;
+import com.lframework.xingyun.sc.dto.stock.ProductStockChangeDto;
 import com.lframework.starter.web.inner.entity.SysUser;
 import com.lframework.starter.web.inner.service.GenerateCodeService;
 import com.lframework.starter.web.inner.service.system.SysUserService;
@@ -365,7 +365,7 @@ public class RetailOutSheetServiceImpl extends
         SubProductStockVo subProductStockVo = new SubProductStockVo();
         subProductStockVo.setProductId(detail.getProductId());
         subProductStockVo.setScId(sheet.getScId());
-        subProductStockVo.setStockNum(detail.getOrderNum());
+        subProductStockVo.setStockNum(BigDecimal.valueOf(detail.getOrderNum()));
         subProductStockVo.setBizId(sheet.getId());
         subProductStockVo.setBizDetailId(detail.getId());
         subProductStockVo.setBizCode(sheet.getCode());
@@ -416,7 +416,7 @@ public class RetailOutSheetServiceImpl extends
           SubProductStockVo subProductStockVo = new SubProductStockVo();
           subProductStockVo.setProductId(newDetail.getProductId());
           subProductStockVo.setScId(sheet.getScId());
-          subProductStockVo.setStockNum(newDetail.getOrderNum());
+          subProductStockVo.setStockNum(BigDecimal.valueOf(newDetail.getOrderNum()));
           subProductStockVo.setBizId(sheet.getId());
           subProductStockVo.setBizDetailId(newDetail.getId());
           subProductStockVo.setBizCode(sheet.getCode());
