+++ xingyun-sc/src/main/java/com/lframework/xingyun/sc/vo/purchase/returned/ReturnProductVo.java
@@ -27,7 +27,7 @@ public class ReturnProductVo implements BaseVo, Serializable {
    * 退货数量
    */
   @ApiModelProperty("退货数量")
-  private Integer returnNum;
+  private BigDecimal returnNum;
 
   /**
    * 备注
